Hoist up prompt toolkit application in test suite
This makes it easier to share with other features that will need
it such as wizards
