+++ tests/functional/autoprompt/test_prompttoolkit.py
@@ -15,11 +15,7 @@ import io
 import os
 
 from prompt_toolkit import Application
-from prompt_toolkit.key_binding.key_processor import KeyPress
-from prompt_toolkit.input import DummyInput
-
 from prompt_toolkit.keys import Keys
-from prompt_toolkit.utils import Event
 
 from awscli.autocomplete.main import create_autocompleter
 from awscli.autocomplete import generator, filters, parser
@@ -31,6 +27,7 @@ from awscli.autoprompt.prompttoolkit import (
 )
 from awscli.autoprompt.history import HistoryDriver
 from awscli.testutils import unittest, random_chars, FileCreator, cd
+from tests import PromptToolkitApplicationStubber as ApplicationStubber
 
 
 def _ec2_only_command_table(command_table, **kwargs):
@@ -55,68 +52,6 @@ class FakeApplication:
 
     def run(self, pre_run):
         pre_run()
-
-
-class FakeInput(DummyInput):
-
-    def fileno(self):
-        return 0
-
-    @property
-    def closed(self):
-        return False
-
-
-class ApplicationStubber:
-    _KEYPRESS = 'KEYPRESS'
-    _ASSERTION = 'ASSERTION'
-
-    def __init__(self, app):
-        self._app = app
-        self._app.input = FakeInput()
-        self._queue = []
-
-    def add_keypress_with_assertion(self, key, assertion):
-        self.add_keypress_action(key)
-        self.add_assertion_action(assertion, key)
-
-    def add_keypress_action(self, key):
-        self._queue.append((self._KEYPRESS, key))
-
-    def add_assertion_action(self, assertion, key=None):
-        self._queue.append((self._ASSERTION, assertion, key))
-
-    def run(self, pre_run=None):
-        key_processor = self._app.key_processor
-        # After each rendering application will run this callback
-        # it takes the next action from the queue and performs it
-        # some key_presses can lead to rerender, after which this callback
-        # will be run again before re-rendering app.invalidated property
-        # set to True.
-        # On exit this callback also run so we need to remove it before exit
-
-        def callback(app):
-            while self._queue and not app.invalidated:
-                action = self._queue.pop(0)
-                if action[0] == self._KEYPRESS:
-                    key_processor.feed(KeyPress(action[1], ''))
-                    key_processor.process_keys()
-                if action[0] == self._ASSERTION:
-                    if not action[1](app):
-                        app.after_render = Event(app, None)
-                        app.exit(
-                            exception=AssertionError(f'Incorrect action on key '
-                                                     f'press "{action[2]}"'))
-                        return
-            if not self._queue:
-                app.after_render = Event(app, None)
-                app.exit()
-
-        self._app.after_render = Event(self._app, callback)
-
-        _stdout = io.TextIOWrapper(io.BytesIO(), encoding="utf-8")
-        self._app.output.stdout = _stdout
-        self._app.run(pre_run=pre_run)
 
 
 class BasicPromptToolkitTest(unittest.TestCase):
